+++ src/test/java/org/mmarini/railways2/model/TrainWaitingForRunTest.java
@@ -41,6 +41,7 @@ import java.util.Optional;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mmarini.railways.Matchers.optionalOf;
 import static org.mmarini.railways2.model.Matchers.locatedAt;
 
 class TrainWaitingForRunTest {
@@ -63,7 +64,7 @@ class TrainWaitingForRunTest {
                 .addEdge(Track.builder("ab"), "a", "b")
                 .addEdge(Track.builder("bc"), "b", "c")
                 .build();
-        status = new StationStatus.Builder(stationMap)
+        status = new StationStatus.Builder(stationMap, 1)
                 .addRoute(Entry::create, "a")
                 .addRoute(Signal::create, "b")
                 .addRoute(Exit::create, "c")
@@ -91,6 +92,6 @@ class TrainWaitingForRunTest {
         Train next = nextOpt.orElseThrow();
         assertEquals(Train.WAITING_FOR_RUN_STATE, next.getState());
         assertEquals(0, next.getSpeed());
-        assertThat(next.getLocation(), locatedAt("ab", "b", 0));
+        assertThat(next.getLocation(), optionalOf(locatedAt("ab", "b", 0)));
     }
 }